feat: add GridViewAdmin and ListViewAdmin components for product management
