+++ index.html
@@ -1,14 +1,57 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/x-icon" href="favicon_io/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>RazorBills</title>
 
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Bakbak+One&family=Outfit:wght@100..900&display=swap" rel="stylesheet">
+    <!-- Open Graph Metadata (Facebook, WhatsApp, LinkedIn, etc.) -->
+    <meta
+      property="og:title"
+      content="RazorBills - Your One-Stop Billing Solution"
+    />
+    <meta
+      property="og:description"
+      content="RazorBills is a comprehensive billing solution that simplifies your invoicing and payment processes. Manage your bills efficiently with our user-friendly platform."
+    />
+    <meta
+      property="og:image"
+      content="https://razorbills.vercel.app/favicon_io/favicon-32x32.jpg"
+    />
+    <meta
+      property="og:logo"
+      content="https://razorbills.vercel.app/favicon_io/favicon-32x32.jpg"
+    />
+    <meta property="og:url" content="https://razorbills.vercel.app" />
+    <meta property="og:type" content="website" />
+    <meta property="og:site_name" content="RazorBills" />
+
+    <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="favicon_io/apple-touch-icon.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="32x32"
+      href="favicon_io/favicon-32x32.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="16x16"
+      href="favicon_io/favicon-16x16.png"
+    />
+    <link rel="manifest" href="favicon_io/site.webmanifest" />
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Bakbak+One&family=Outfit:wght@100..900&display=swap"
+      rel="stylesheet"
+    />
   </head>
   <body>
     <div id="root"></div>
